:sparkles: add new Features

+++ src/main/java/com/github/modules/sys/controller/SysUserController.java
@@ -1,5 +1,6 @@
 package com.github.modules.sys.controller;
 
+import com.github.common.annotation.SysLog;
 import com.github.common.constant.SysConstant;
 import com.github.common.exception.SHException;
 import com.github.common.utils.ApiResponse;
@@ -7,6 +8,8 @@ import com.github.common.utils.PageUtils;
 import com.github.common.validator.ValidatorUtils;
 import com.github.common.validator.group.AddGroup;
 import com.github.common.validator.group.UpdateGroup;
+import com.github.modules.base.form.PageForm;
+import com.github.modules.data.entity.RuleEntity;
 import com.github.modules.sys.entity.SysUserEntity;
 import com.github.modules.sys.service.SysUserService;
 import org.apache.commons.lang.ArrayUtils;
@@ -27,7 +30,7 @@ import java.util.Objects;
  */
 @Controller
 @RequestMapping("/sys/user")
-public class SysUserController {
+public class SysUserController extends AbstractController {
 
     @Autowired
     private SysUserService sysUserService;
@@ -46,12 +49,12 @@ public class SysUserController {
     /**
      * 所有用户列表
      *
-     * @param params 请求参数
+     * @param pageForm 请求参数
      */
     @GetMapping("/list")
     @ResponseBody
-    public ApiResponse list(@RequestParam Map<String, String> params) {
-        PageUtils pageBean = sysUserService.findPage(params);
+    public ApiResponse list(PageForm pageForm) {
+        PageUtils pageBean = sysUserService.findPage(pageForm);
         return ApiResponse.ofSuccess().put("pageBean", pageBean);
     }
 
@@ -61,9 +64,11 @@ public class SysUserController {
     @GetMapping("/verify")
     @ResponseBody
     public ApiResponse verify(@RequestParam String username) {
-        if (sysUserService.findByUsername(username) == null) {
-            return null;
+        SysUserEntity sysUserEntity = sysUserService.findByName(username);
+        if (sysUserEntity == null) {
+            return ApiResponse.ofSuccess();
         }
+
         return ApiResponse.ofFail("用户名已存在");
 
     }
@@ -71,11 +76,13 @@ public class SysUserController {
     /**
      * 保存用户
      */
+    @SysLog("保存用户")
     @PostMapping("/save")
     @ResponseBody
     public ApiResponse save(SysUserEntity userEntity) {
         ValidatorUtils.validateEntity(userEntity, AddGroup.class);
 
+        userEntity.setCreateUserId(getUserId());
         sysUserService.save(userEntity);
 
         return ApiResponse.ofSuccess();
@@ -84,15 +91,16 @@ public class SysUserController {
     /**
      * 用户信息
      */
-    @GetMapping("/info/{userId:[1-9]+}")
+    @GetMapping("/info/{userId:[0-9]+}")
     public String info(@PathVariable Long userId, Model model){
-        model.addAttribute("user", sysUserService.findByUserId(userId));
+        model.addAttribute("user", sysUserService.findById(userId));
         return "admin/sys/user/userInfo";
     }
 
     /**
      * 修改用户
      */
+    @SysLog("修改用户")
     @PutMapping("/update")
     @ResponseBody
     public ApiResponse update(@ModelAttribute("user") SysUserEntity sysUserEntity){
@@ -109,6 +117,7 @@ public class SysUserController {
 
         ValidatorUtils.validateEntity(sysUserEntity, UpdateGroup.class);
 
+        sysUserEntity.setCreateUserId(getUserId());
         sysUserService.update(sysUserEntity);
 
         return ApiResponse.ofSuccess();
@@ -117,11 +126,16 @@ public class SysUserController {
     /**
      * 删除用户
      */
+    @SysLog("删除用户")
     @DeleteMapping("/delete")
     @ResponseBody
     public ApiResponse delete(@RequestParam("selectIds") Long[] userIds) {
         if (ArrayUtils.contains(userIds, SysConstant.SUPER_ADMIN)) {
             return ApiResponse.ofFail("超级管理员不能删除");
+        }
+
+        if(ArrayUtils.contains(userIds, getUserId())){
+            return ApiResponse.ofFail("当前用户不能删除");
         }
 
         sysUserService.deleteBatch(userIds);
@@ -132,7 +146,8 @@ public class SysUserController {
     /**
      * 重置密码
      */
-    @PutMapping("/pwd/{userId:[1-9]+}")
+    @SysLog("重置密码")
+    @PutMapping("/pwd/{userId:[0-9]+}")
     @ResponseBody
     public ApiResponse reset(@PathVariable Long userId) {
         sysUserService.reset(userId);
@@ -142,7 +157,7 @@ public class SysUserController {
     @ModelAttribute
     private void customModelAttribute(@RequestParam(value = "userId", required = false) Long userId, Model model) {
         if (userId != null) {
-            SysUserEntity userEntity = sysUserService.findByUserId(userId);
+            SysUserEntity userEntity = sysUserService.findById(userId);
             model.addAttribute("user", userEntity);
         }
     }
